added discount loading charges

+++ database/migrations/2023_07_01_052030_create_estimates_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('customer_id');
             $table->string('estimate_id');
-            $table->float('total_amount');
+            $table->double('total_amount');
             $table->timestamps();
         });
     }
